# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Supermarket API Documentation",
        "description": "Documentation for Supermarket Departments API",
        "contact": {
            "name": "API Support",
            "email": "contact@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/v1/stock/statistics": {
            "get": {
                "tags": [
                    "Stock"
                ],
                "summary": "Get stock statistics",
                "operationId": "ab1cb76e8940b84c71ef0eabf199f4ef",
                "responses": {
                    "200": {
                        "description": "Stock statistics",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "out_of_stock_count": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "critical_stock_count": {
                                                    "type": "integer",
                                                    "example": 12
                                                },
                                                "department_average_stock": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "department": {
                                                                "type": "string",
                                                                "example": "Electronics"
                                                            },
                                                            "average_stock": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 45.5
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "department_product_counts": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "department": {
                                                                "type": "string",
                                                                "example": "Electronics"
                                                            },
                                                            "total_products": {
                                                                "type": "integer",
                                                                "example": 25
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}